clk: qcom: clk-hfpll: Implement mechanism to safely park the PLL
The HFPLL may get enabled with the previous L register programming,
which would raise it at the previous frequency.
If the previous frequency is higher than the one that we want to
set, the VDD vote will be lower than normal, resulting in PLL
undervolting and locking failures.
This is a rather rare condition from which we can indeed recover
later and automatically but if we handle it like this we will
overall lose less time.